add a plugin option to disable 3d systems
this allow to use this plugin in a headless bevy app, for example with
MinimalPlugins, and so without bevy_render.

changes:
- add a optional_3d boolean to the plugin struct
- impl Default for the plugin struct, enabling by default the 3d systems
- update all the examples to use the Default plugin impl
- add the headless example, it just print to stdout the events from the
basic example
- replace most of the u32 used interally to track time by Duration. this
fix hidden dangerous bugs created by either float inprecision or just
wrong code because of the non-fixed unit of value. Duration ensure the
duration are correctly calculated and used together.
the primary bug that lead to this fix was running the app with
bevy::app::ScheduleRunnerPlugin (from MinimalPlugins) which mean the app
refreshed as fast as it can, instead of the screen refresh rate. this
lead the Time interval to be microseconds wide instead of milliseconds.
inaccurary in the durations calculations lead to the plugin not working
(animations werent advancing).
- fix the pause_resume test values. now the frame update correctly
without loosing durations